Rearrange devDeps to match Yarn expectations
It appears that the CI functions due to a fluke in how the node modules are resolved. This changes fixes that, providing the devDeps a package(Workspace) needs, rather than relying on it being able to see top-level package.json devDeps.

Add "yarn run ci" to locally run the same thigns as CI will run. Tidy a couple others (explicit "run" along the way.