Test harness for React TS output
This builds and runs, but requires a ts-ignore. Mitosis React TSX output currently expects to be in "source" rather than node_modules, so it doesn't compile the TS(X) and doesn't include .d.ts. This could be enhanced in the future to facilitate easy use of Mitosis in one project of a monorepo, with output consume by other projects easily.