better support both forms of state and functions in Mitosis (#480)
* better support both forms of state and functions in Mitosis

currently, we have a "useState" hook that supports two different formats,
based on how it is used. this changes the default name
for the object format to useStore({}) and keeps useState()
to be used just for the react format. this also fixes types accordingly
and allows you to intermix both types of hooks and functions as you please

so now you can do:

```ts
export default function MyCompnent(props) {
  const state = useStore({
    myProperty: 'foo',
    myMethod() { /* ... */ },
    get myGetter() { /* ... */ },
 })
 function myFunction() { /* ... */ }
 const [foo, setFoo] = useState('foobar')

 return <>/*...*/</>
}
```

pros:
 - one name does one thing
 - avoid confusing of react's useState semantics and (part of) ours
 - this is backwards compatible, you can still use useState({...}) for now so you don't have to refactor old code
 - fixes/improves TS types

 cons:
  - `const state = useStore(...)` may cause confusion. state object vs state hooks

* fix

* update eslint plugin

* typo

* prettier

* refactor(eslint): reuse hooks names

* refactor(core): reuse name const

* fix import

Co-authored-by: Sami Jaber <jabersami@gmail.com>